backend/graphrag/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
neo4j==5.19.0
qdrant-client==1.9.1
numpy==1.26.4
transformers==4.39.3
sentence-transformers==2.7.0
httpx==0.27.0
openai==1.30.1
anthropic==0.28.0
redis==5.0.4
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.1
PyJWT>=2.8.0
